requirements.txt
python-dotenv>=1.0.0
netifaces>=0.11.0
requests>=2.31.0
urllib3>=2.0.0
pika>=1.3.2
